+++ core/docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * Implement the `StartExternalContainerRegistration` and `FinishExternalContainerRegistration` endpoints
+* Always bind the API container to a host machine port
 
 ### Changes
 * Upgraded to container-engine-lib 0.4.0, which replaces the long list of `CreateAndStartContainer` args with a builder